requirements.txt
hugchat == 0.3.7
python-dotenv == 1.0.0
typer[all] == 0.9.0
GitPython == 3.1.40
